+++ google/cloud/sql/connector/connector.py
@@ -26,6 +26,7 @@ import google.cloud.sql.connector.pg8000 as pg8000
 import google.cloud.sql.connector.pytds as pytds
 import google.cloud.sql.connector.asyncpg as asyncpg
 from google.cloud.sql.connector.utils import generate_keys
+from google.cloud.sql.connector.exceptions import ConnectorLoopError
 from google.auth.credentials import Credentials
 from threading import Thread
 from typing import Any, Dict, Optional, Type
@@ -34,16 +35,6 @@ from functools import partial
 logger = logging.getLogger(name=__name__)
 
 ASYNC_DRIVERS = ["asyncpg", "aiomysql"]
-
-
-class ConnectorLoopError(Exception):
-    """
-    Raised when Connector.connect is called with Connector._loop
-        in an invalid state (event loop in current thread).
-    """
-
-    def __init__(self, *args: Any) -> None:
-        super(ConnectorLoopError, self).__init__(self, *args)
 
 
 class Connector:
@@ -67,10 +58,22 @@ class Connector:
         Credentials object used to authenticate connections to Cloud SQL server.
         If not specified, Application Default Credentials are used.
 
+    :type quota_project: str
+    :param quota_project
+        The Project ID for an existing Google Cloud project. The project specified
+        is used for quota and billing purposes. If not specified, defaults to
+        project sourced from environment.
+
     :type loop: asyncio.AbstractEventLoop
     :param loop
         Event loop to run asyncio tasks, if not specified, defaults to
         creating new event loop on background thread.
+
+    :type sqladmin_api_endpoint: str
+    :param sqladmin_api_endpoint:
+        Base URL to use when calling the Cloud SQL Admin API endpoint.
+        Defaults to "https://sqladmin.googleapis.com", this argument should
+        only be used in development.
     """
 
     def __init__(
@@ -80,6 +83,8 @@ class Connector:
         timeout: int = 30,
         credentials: Optional[Credentials] = None,
         loop: asyncio.AbstractEventLoop = None,
+        quota_project: Optional[str] = None,
+        sqladmin_api_endpoint: str = "https://sqladmin.googleapis.com",
     ) -> None:
         # if event loop is given, use for background tasks
         if loop:
@@ -101,6 +106,8 @@ class Connector:
         self._timeout = timeout
         self._enable_iam_auth = enable_iam_auth
         self._ip_type = ip_type
+        self._quota_project = quota_project
+        self._sqladmin_api_endpoint = sqladmin_api_endpoint
         self._credentials = credentials
 
     def connect(
@@ -196,6 +203,8 @@ class Connector:
                 self._loop,
                 self._credentials,
                 enable_iam_auth,
+                self._quota_project,
+                self._sqladmin_api_endpoint,
             )
             self._instances[instance_connection_string] = instance
 
@@ -213,14 +222,7 @@ class Connector:
         except KeyError:
             raise KeyError(f"Driver '{driver}' is not supported.")
 
-        if "ip_types" in kwargs:
-            ip_type = kwargs.pop("ip_types")
-            logger.warning(
-                "Deprecation Warning: Parameter `ip_types` is deprecated and may be removed"
-                " in a future release. Please use `ip_type` instead."
-            )
-        else:
-            ip_type = kwargs.pop("ip_type", self._ip_type)
+        ip_type = kwargs.pop("ip_type", self._ip_type)
         timeout = kwargs.pop("timeout", self._timeout)
         if "connect_timeout" in kwargs:
             timeout = kwargs.pop("connect_timeout")
@@ -248,10 +250,10 @@ class Connector:
             return await asyncio.wait_for(get_connection(), timeout)
         except asyncio.TimeoutError:
             raise TimeoutError(f"Connection timed out after {timeout}s")
-        except Exception as e:
+        except Exception:
             # with any other exception, we attempt a force refresh, then throw the error
             instance.force_refresh()
-            raise (e)
+            raise
 
     def __enter__(self) -> Any:
         """Enter context manager by returning Connector object"""
@@ -286,6 +288,12 @@ class Connector:
         )
         # Will attempt to safely shut down tasks for 5s
         close_future.result(timeout=5)
+        # if background thread exists for Connector, clean it up
+        if self._thread:
+            # stop event loop running in background thread
+            self._loop.call_soon_threadsafe(self._loop.stop)
+            # wait for thread to finish closing (i.e. loop to stop)
+            self._thread.join()
 
     async def close_async(self) -> None:
         """Helper function to cancel Instances' tasks